% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serology_analysis.R
\name{calculate_unadjusted_gam}
\alias{calculate_unadjusted_gam}
\title{Perform an unadjusted gam fit for each variable in a formula given a dataframe}
\usage{
calculate_unadjusted_gam(model, formula)
}
\arguments{
\item{model}{input dataframe to be fitted to the GAM model}

\item{formula}{a full formula for a (possible) multivariate game}
}
\description{
Perform an unadjusted gam fit for each variable in a formula given a dataframe
}
