% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serology_analysis.R
\name{code_vars}
\alias{code_vars}
\title{Helper function to code the key variables used in the serology analysis
e.g. making factor variables}
\usage{
code_vars(model)
}
\arguments{
\item{model}{}
}
\value{
modified model
}
\description{
Helper function to code the key variables used in the serology analysis
e.g. making factor variables
}
