% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serology_analysis.R
\name{get_coeff_name_map}
\alias{get_coeff_name_map}
\title{Get a map between the coefficient name and
\itemize{
\item the original variable it came from
\item the level it corresponds to
This function is for use with factor levels in GAM Fits
}}
\usage{
get_coeff_name_map(gamFit)
}
\arguments{
\item{gamFit}{GAM fit object from mgcv}
}
\description{
Get a map between the coefficient name and
\itemize{
\item the original variable it came from
\item the level it corresponds to
This function is for use with factor levels in GAM Fits
}
}
