% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_data_dictionary}
\alias{get_data_dictionary}
\title{Get the Data Dictionary}
\usage{
get_data_dictionary()
}
\value{
the full EAVE data dictionary as a data.frame
}
\description{
Return the full data dictionary
}
