% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_description}
\alias{get_description}
\title{Get the Description}
\usage{
get_description(var)
}
\arguments{
\item{the}{variable name}
}
\value{
a description (or multiple, if multiple found) of the variable
}
\description{
Pass a variable name (as a string) and get the description(s) found in
the data dictionary returned
}
