% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_label}
\alias{get_label}
\title{Get the Label}
\usage{
get_label(var)
}
\arguments{
\item{the}{variable name}
}
\value{
a label (or multiple, if multiple found) of the variable
}
\description{
Pass a variable name (as a string) and get the label(s) found in
the data dictionary returned
}
