% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serology_analysis.R
\name{get_modelA_df}
\alias{get_modelA_df}
\title{Simple function to get the dataframe for modelA given a cohort dataframe}
\usage{
get_modelA_df(df)
}
\arguments{
\item{df}{input data for the cohort}
}
\value{
cleaned and mutated version of the data for analysis
}
\description{
Simple function to get the dataframe for modelA given a cohort dataframe
}
