% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serology_analysis.R
\name{get_or_from_gam}
\alias{get_or_from_gam}
\title{Extract ORs from a GAM fit for the parametric variables}
\usage{
get_or_from_gam(mod, level = 0.95)
}
\arguments{
\item{mod}{the GAM fit object from mgcv}
}
\description{
Extract ORs from a GAM fit for the parametric variables
}
