% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serology_analysis.R
\name{get_pterm_or_from_gam}
\alias{get_pterm_or_from_gam}
\title{..... REDUNDANT FUNCTION......
Extract the parametric terms from a GAM fit
.. these are all variables that are not splines or interactions..}
\usage{
get_pterm_or_from_gam(fit, model)
}
\arguments{
\item{fit}{the GAM fit object}

\item{model}{the input dataframe used for the model}
}
\description{
..... REDUNDANT FUNCTION......
Extract the parametric terms from a GAM fit
.. these are all variables that are not splines or interactions..
}
