% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_var}
\alias{get_var}
\title{Look up a variable in the data dictionary}
\usage{
get_var(var)
}
\arguments{
\item{the}{variable name}
}
\value{
a data.frame of information found about the variable input
}
\description{
Return all information found about this variable by name
}
