% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serology_analysis.R
\name{perform_gam}
\alias{perform_gam}
\title{Perform a GAM fit of a model given a formula
in the serology GAM analys(es) we always use binomial logistic regression
.. so this is a simple convienence function}
\usage{
perform_gam(model, formula)
}
\arguments{
\item{model}{input dataframe to be used}

\item{formula}{formula string (e.g. y ~ x + z)}
}
\description{
Perform a GAM fit of a model given a formula
in the serology GAM analys(es) we always use binomial logistic regression
.. so this is a simple convienence function
}
