% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{serology_outcome_analysis.create_hosp_death_dataframe}
\alias{serology_outcome_analysis.create_hosp_death_dataframe}
\title{Function to help create the dataframe needed for the hosp/death analysis
\itemize{
\item given an already built cohort dataframe, we build a new dataframe to link with hospitalisation/death outcomes
}}
\usage{
serology_outcome_analysis.create_hosp_death_dataframe(cohort, eave.data)
}
\arguments{
\item{cohort}{initial dataframe built for a cohort containing the serology data}

\item{eave.data}{all EAVE-II recommended datasets}
}
\value{
analysis data frame
}
\description{
Function to help create the dataframe needed for the hosp/death analysis
\itemize{
\item given an already built cohort dataframe, we build a new dataframe to link with hospitalisation/death outcomes
}
}
