% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{serology_outcome_analysis.prepare_dataframe}
\alias{serology_outcome_analysis.prepare_dataframe}
\title{Generic to create a dataframe for a cox regression analysis
\itemize{
\item given an already built cohort dataframe, we build a new dataframe to link with hospitalisation/death outcomes
}}
\usage{
serology_outcome_analysis.prepare_dataframe(df)
}
\value{
analysis data frame
}
\description{
Generic to create a dataframe for a cox regression analysis
\itemize{
\item given an already built cohort dataframe, we build a new dataframe to link with hospitalisation/death outcomes
}
}
