% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serology_analysis.R
\name{calculate_product_category}
\alias{calculate_product_category}
\title{Hacky bit of code to classify product-stage (vaccine dose) combinations}
\usage{
calculate_product_category(stage, d1, d2, d3, d4)
}
\description{
Hacky bit of code to classify product-stage (vaccine dose) combinations
}
