% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serology_analysis.R
\name{get_vars_to_use}
\alias{get_vars_to_use}
\title{Work out which variables are good enough to be used for analysis
based upon how good the p-value is for them when used in a univariate GLM fit with the outcome}
\usage{
get_vars_to_use(model, outcome_var, thres = 0.1)
}
\description{
Work out which variables are good enough to be used for analysis
based upon how good the p-value is for them when used in a univariate GLM fit with the outcome
}
